change project structure to feature-based
