+++ src/app/about/about.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+
+
+// Only because used Material's Sidenav in AppComponent which expects
+// BrowserModule eagerly(non-lazy) loaded
+@NgModule({
+  declarations: [],
+  imports: [],
+  providers: [],
+  bootstrap: []
+})
+export class AboutModule { }
