--- src/app/app.component.css
@@ -1,1 +1,0 @@
-
