--- src/app/app.component.html
@@ -1,22 +1,0 @@
-
-<mat-sidenav-container autosize>
-  <mat-sidenav class="sidenav mat-elevation-z4" mode="side" opened>
-    <mat-nav-list>
-      <a mat-list-item>
-        <button mat-raised-button color="primary" routerLink="prodview">
-          ProductView
-        </button>
-      </a>
-      <a mat-list-item>
-        <button mat-raised-button color="warn" routerLink="abtview">
-          About
-        </button>
-      </a>
-    </mat-nav-list>
-  </mat-sidenav>
-  <mat-sidenav-content>
-    <div style="height: 100vh;">
-      <router-outlet></router-outlet>
-    </div>
-  </mat-sidenav-content>
-</mat-sidenav-container>
