--- src/app/components/about-view/about-view.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-about-view',
-  templateUrl: './about-view.component.html',
-  styleUrls: ['./about-view.component.css']
-})
-export class AboutViewComponent {
-
-}
