--- src/app/components/product-view/product-view.component.css
@@ -1,16 +1,0 @@
-.bf-container {
-  padding: 16px;
-}
-
-#c1 {
-  float: left;
-}
-
-#c2 {
-  float: right;
-}
-
-.table-container {
-  padding: 16px;
-}
-
