--- src/app/components/product-view/product-view.component.html
@@ -1,49 +1,0 @@
-<div class="bf-container">
-  <button mat-raised-button id="c1"> Add </button>
-  <mat-form-field id="c2">
-    <input (keyup)="applyFilter($any($event.target).value)" matInput placeholder="Filter">
-  </mat-form-field>
-</div>
-
-<div class="table-container">
-  <table mat-table [dataSource]="dataSource" matSort class="mat-elevation-z8">
-    <!-- Position Column -->
-    <ng-container matColumnDef="name">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header> Name </th>
-      <td mat-cell *matCellDef="let element"> {{element.name}} </td>
-    </ng-container>
-    <!-- Position Column -->
-    <ng-container matColumnDef="manufacturer.name">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header> Manufacturer </th>
-      <td mat-cell *matCellDef="let element"> {{element.manufacturer.name}} </td>
-    </ng-container>
-    <!-- Position Column -->
-    <ng-container matColumnDef="price">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header> Price </th>
-      <td mat-cell *matCellDef="let element"> {{element.price}}€ </td>
-    </ng-container>
-    <!-- Position Column -->
-    <ng-container matColumnDef="expiryDate">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header> Expiry Date </th>
-      <td mat-cell *matCellDef="let element"> {{element.expiryDate | date:'shortDate'}} </td>
-    </ng-container>
-    <!-- Position Column -->
-    <ng-container matColumnDef="Edit">
-      <th mat-header-cell *matHeaderCellDef> </th>
-      <td mat-cell *matCellDef="let element"> <button mat-raised-button color="primary"> Edit </button> </td>
-    </ng-container>
-    <!-- Position Column -->
-    <ng-container matColumnDef="Delete">
-      <th mat-header-cell *matHeaderCellDef> </th>
-      <td mat-cell *matCellDef="let element"> <button mat-raised-button color="warn"> Delete </button> </td>
-    </ng-container>
-
-    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-    <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-  </table>
-  <mat-paginator #paginator class="mat-elevation-z8"
-                 [pageSizeOptions]="[5, 10, 25, 100]"
-                 showFirstLastButtons
-                 aria-label="Select page">
-  </mat-paginator>
-</div>
