+++ src/app/core/components/app/app.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-root',
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.css']
+  // dependencies(imports) are in app.module
+})
+export class AppComponent {
+  title = 'PharmacyFront';
+}
