+++ src/app/core/models/imanufacturer.ts
@@ -1,0 +1,4 @@
+export interface IManufacturer {
+  id: string;
+  name: string;
+}
