+++ src/app/core/models/iproduct.ts
@@ -1,0 +1,9 @@
+import {IManufacturer} from "./imanufacturer";
+
+export interface IProduct {
+  id: string;
+  name: string;
+  manufacturer: IManufacturer;
+  price: number;
+  expiryDate: Date;
+}
