--- src/app/interfaces/imanufacturer.ts
@@ -1,4 +1,0 @@
-export interface IManufacturer {
-  id: string;
-  name: string;
-}
