--- src/app/interfaces/iproduct.ts
@@ -1,9 +1,0 @@
-import {IManufacturer} from "./imanufacturer";
-
-export interface IProduct {
-  id: string;
-  name: string;
-  manufacturer: IManufacturer;
-  price: number;
-  expiryDate: Date;
-}
