+++ src/app/product/components/product-view/product-view.component.css
@@ -1,0 +1,16 @@
+.bf-container {
+  padding: 16px;
+}
+
+#c1 {
+  float: left;
+}
+
+#c2 {
+  float: right;
+}
+
+.table-container {
+  padding: 16px;
+}
+
