+++ src/app/product/product-routing.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+const productRoutes: Routes = [
+
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(productRoutes)],
+  exports: [RouterModule]
+})
+export class ProductRoutingModule { }
