--- src/app/services/product.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ProductService } from './product.service';
-
-describe('ProductService', () => {
-  let service: ProductService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ProductService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
