+++ src/main.ts
@@ -1,8 +1,21 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from './app/app.module';
+import {bootstrapApplication} from "@angular/platform-browser";
+import {AppComponent} from "./app/core/components/app/app.component";
+import {provideRouter} from "@angular/router";
+// import routes from "./app/app.routes";
+import {AppModule} from "./app/core/app.module";
 
 
+// We must use non-standalone bootstrapApp because of Material Sidenav
+// which expects BrowserModule to be eagerly(non-lazy) loaded
 platformBrowserDynamic().bootstrapModule(AppModule)
   .catch(err => console.error(err));
+
+/*
+bootstrapApplication(AppComponent,{
+  providers:[provideRouter(routes)]
+}).catch(err => console.error(err));
+*/
+
